pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tool-redmine-attachment-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "playwright>=1.53.0",
  "python-dotenv>=1.1.1",
  "requests>=2.32.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
